+++ src/libbluray/bdnav/index_parse.c
@@ -199,7 +199,7 @@ INDX_ROOT *indx_parse(const char *file_name)
 
     /* if failed, try backup file */
     if (!indx) {
-        int   len    = strlen(file_name);
+        size_t len   = strlen(file_name);
         char *backup = malloc(len + 8);
 
         strcpy(backup, file_name);
